color-cifar10-shift/requirements.txt
torch>=2.1
torchvision
laplace-torch
datasets
numpy
pillow
matplotlib
scikit-learn
tqdm
